Update 3.4
Reconfiguration of CameraFun
